requirements.txt
asgiref==3.8.1
Django==4.2.14
django-telegram-login==0.2.3
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
